refactor(polaris): separate config package, removed files
