+++ src/scanners/polaris/examples/scanner-config.yaml
@@ -2,12 +2,12 @@ scanner:
   id: 8fc1b601-d5a2-466d-8236-c747c1dc02a2
   periodicity: "0 2 * * *"
   heartbeat-periodicity: 86400
-  version: 0.1.0
+  version: 0.1.2
 polaris:
   configPath: ./examples/polaris-config.yaml
   version: 0.6.0
 blobStorageType: azure-blob-storage
-azureBlobConfig:
+azureBlob:
   storageBaseUrl: https://BLOB_NAME.blob.core.windows.net/CONTAINER_NAME
   sasToken: insert-sas-token-here
 logFormat: plain-text
